feat(concurrency): implement proper Redis distributed locking for account operations  Replaced incorrect cache-based locking mechanism with a proper Redis distributed lock using IConnectionMultiplexer.  - Implemented atomic lock acquisition using SET NX with expiration - Ensured safe lock release using ownership validation - Fixed Redis connection setup and configuration - Enabled Redis via Docker for local development environment - Removed reliance on IDistributedCache for concurrency control  This change ensures correct mutual exclusion for critical operations such as withdrawals, preventing race conditions and improving data consistency in a financial context.

+++ BusinessLayer/Service/IService/IAccountService.cs
@@ -19,7 +19,7 @@ namespace BusinessLayer.Service.IService
 
         Task WithdrawAsync(string accountNumber, decimal balance);
 
-        Task<double> GetBalanceAsync(string accountNumber);
+        Task<decimal> GetBalanceAsync(string accountNumber);
 
         Task<bool> TransferAmountAsync(string senderId, string receiverId, decimal amount);
     }
